+++ fltk-sys/src/misc.rs
@@ -314,6 +314,15 @@ extern "C" {
     pub fn Fl_Spinner_callback(self_: *const Fl_Spinner) -> Fl_Callback;
 }
 extern "C" {
+    pub fn Fl_Spinner_set_deletion_callback(
+        self_: *mut Fl_Spinner,
+        arg1: ::std::option::Option<
+            unsafe extern "C" fn(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
     pub fn Fl_Spinner_minimum(arg1: *mut Fl_Spinner) -> f64;
 }
 extern "C" {
@@ -665,6 +674,15 @@ extern "C" {
 extern "C" {
     pub fn Fl_Clock_callback(self_: *const Fl_Clock) -> Fl_Callback;
 }
+extern "C" {
+    pub fn Fl_Clock_set_deletion_callback(
+        self_: *mut Fl_Clock,
+        arg1: ::std::option::Option<
+            unsafe extern "C" fn(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Chart {
@@ -961,6 +979,15 @@ extern "C" {
     pub fn Fl_Chart_callback(self_: *const Fl_Chart) -> Fl_Callback;
 }
 extern "C" {
+    pub fn Fl_Chart_set_deletion_callback(
+        self_: *mut Fl_Chart,
+        arg1: ::std::option::Option<
+            unsafe extern "C" fn(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
     pub fn Fl_Chart_clear(self_: *mut Fl_Chart);
 }
 extern "C" {
@@ -1024,6 +1051,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Chart_make_autosize(self_: *mut Fl_Chart, n: ::std::os::raw::c_int);
+}
+extern "C" {
+    pub fn Fl_Chart_bounds(self_: *mut Fl_Chart, a: *mut f64, b: *mut f64);
 }
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
@@ -1321,6 +1351,15 @@ extern "C" {
     pub fn Fl_Progress_callback(self_: *const Fl_Progress) -> Fl_Callback;
 }
 extern "C" {
+    pub fn Fl_Progress_set_deletion_callback(
+        self_: *mut Fl_Progress,
+        arg1: ::std::option::Option<
+            unsafe extern "C" fn(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
     pub fn Fl_Progress_minimum(arg1: *mut Fl_Progress) -> f64;
 }
 extern "C" {
@@ -1366,9 +1405,6 @@ extern "C" {
     pub fn Fl_Tooltip_disable();
 }
 extern "C" {
-    pub static mut Fl_Tooltip_enter: ::std::option::Option<unsafe extern "C" fn(w: *mut Fl_Widget)>;
-}
-extern "C" {
     pub fn Fl_Tooltip_enter_area(
         w: *mut Fl_Widget,
         X: ::std::os::raw::c_int,
@@ -1377,9 +1413,6 @@ extern "C" {
         H: ::std::os::raw::c_int,
         tip: *const ::std::os::raw::c_char,
     );
-}
-extern "C" {
-    pub static mut Fl_Tooltip_exit: ::std::option::Option<unsafe extern "C" fn(w: *mut Fl_Widget)>;
 }
 extern "C" {
     pub fn Fl_Tooltip_current_widget() -> *mut ::std::os::raw::c_void;
@@ -1729,6 +1762,15 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Help_View_callback(self_: *const Fl_Help_View) -> Fl_Callback;
+}
+extern "C" {
+    pub fn Fl_Help_View_set_deletion_callback(
+        self_: *mut Fl_Help_View,
+        arg1: ::std::option::Option<
+            unsafe extern "C" fn(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
 }
 extern "C" {
     pub fn Fl_Help_View_directory(self_: *const Fl_Help_View) -> *const ::std::os::raw::c_char;
@@ -2124,6 +2166,15 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Input_Choice_callback(self_: *const Fl_Input_Choice) -> Fl_Callback;
+}
+extern "C" {
+    pub fn Fl_Input_Choice_set_deletion_callback(
+        self_: *mut Fl_Input_Choice,
+        arg1: ::std::option::Option<
+            unsafe extern "C" fn(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
 }
 extern "C" {
     pub fn Fl_Input_Choice_set_down_box(self_: *mut Fl_Input_Choice, box_: ::std::os::raw::c_int);