+++ fltk-sys/src/window.rs
@@ -321,6 +321,15 @@ extern "C" {
     pub fn Fl_Window_callback(self_: *const Fl_Window) -> Fl_Callback;
 }
 extern "C" {
+    pub fn Fl_Window_set_deletion_callback(
+        self_: *mut Fl_Window,
+        arg1: ::std::option::Option<
+            unsafe extern "C" fn(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
     pub fn Fl_Window_begin(self_: *mut Fl_Window);
 }
 extern "C" {
@@ -511,6 +520,13 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Window_set_icon_label(arg1: *mut Fl_Window, arg2: *const ::std::os::raw::c_char);
+}
+extern "C" {
+    pub fn Fl_Window_set_icons(
+        w: *mut Fl_Window,
+        images: *mut *const ::std::os::raw::c_void,
+        length: ::std::os::raw::c_int,
+    );
 }
 extern "C" {
     pub fn Fl_Window_find_by_handle(handle: *mut ::std::os::raw::c_void) -> *mut Fl_Window;
@@ -872,6 +888,15 @@ extern "C" {
     pub fn Fl_Single_Window_callback(self_: *const Fl_Single_Window) -> Fl_Callback;
 }
 extern "C" {
+    pub fn Fl_Single_Window_set_deletion_callback(
+        self_: *mut Fl_Single_Window,
+        arg1: ::std::option::Option<
+            unsafe extern "C" fn(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
     pub fn Fl_Single_Window_begin(self_: *mut Fl_Single_Window);
 }
 extern "C" {
@@ -1108,6 +1133,13 @@ extern "C" {
     pub fn Fl_Single_Window_set_icon_label(
         arg1: *mut Fl_Single_Window,
         arg2: *const ::std::os::raw::c_char,
+    );
+}
+extern "C" {
+    pub fn Fl_Single_Window_set_icons(
+        w: *mut Fl_Single_Window,
+        images: *mut *const ::std::os::raw::c_void,
+        length: ::std::os::raw::c_int,
     );
 }
 #[repr(C)]
@@ -1448,6 +1480,15 @@ extern "C" {
     pub fn Fl_Double_Window_callback(self_: *const Fl_Double_Window) -> Fl_Callback;
 }
 extern "C" {
+    pub fn Fl_Double_Window_set_deletion_callback(
+        self_: *mut Fl_Double_Window,
+        arg1: ::std::option::Option<
+            unsafe extern "C" fn(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
     pub fn Fl_Double_Window_flush(w: *mut Fl_Double_Window);
 }
 extern "C" {
@@ -1687,6 +1728,13 @@ extern "C" {
     pub fn Fl_Double_Window_set_icon_label(
         arg1: *mut Fl_Double_Window,
         arg2: *const ::std::os::raw::c_char,
+    );
+}
+extern "C" {
+    pub fn Fl_Double_Window_set_icons(
+        w: *mut Fl_Double_Window,
+        images: *mut *const ::std::os::raw::c_void,
+        length: ::std::os::raw::c_int,
     );
 }
 #[repr(C)]
@@ -2003,6 +2051,15 @@ extern "C" {
     pub fn Fl_Menu_Window_callback(self_: *const Fl_Menu_Window) -> Fl_Callback;
 }
 extern "C" {
+    pub fn Fl_Menu_Window_set_deletion_callback(
+        self_: *mut Fl_Menu_Window,
+        arg1: ::std::option::Option<
+            unsafe extern "C" fn(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
     pub fn Fl_Menu_Window_begin(self_: *mut Fl_Menu_Window);
 }
 extern "C" {
@@ -2205,6 +2262,13 @@ extern "C" {
     pub fn Fl_Menu_Window_set_icon_label(
         arg1: *mut Fl_Menu_Window,
         arg2: *const ::std::os::raw::c_char,
+    );
+}
+extern "C" {
+    pub fn Fl_Menu_Window_set_icons(
+        w: *mut Fl_Menu_Window,
+        images: *mut *const ::std::os::raw::c_void,
+        length: ::std::os::raw::c_int,
     );
 }
 #[repr(C)]
@@ -2563,6 +2627,15 @@ extern "C" {
     pub fn Fl_Overlay_Window_callback(self_: *const Fl_Overlay_Window) -> Fl_Callback;
 }
 extern "C" {
+    pub fn Fl_Overlay_Window_set_deletion_callback(
+        self_: *mut Fl_Overlay_Window,
+        arg1: ::std::option::Option<
+            unsafe extern "C" fn(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
     pub fn Fl_Overlay_Window_begin(self_: *mut Fl_Overlay_Window);
 }
 extern "C" {
@@ -2825,6 +2898,13 @@ extern "C" {
     pub fn Fl_Overlay_Window_set_icon_label(
         arg1: *mut Fl_Overlay_Window,
         arg2: *const ::std::os::raw::c_char,
+    );
+}
+extern "C" {
+    pub fn Fl_Overlay_Window_set_icons(
+        w: *mut Fl_Overlay_Window,
+        images: *mut *const ::std::os::raw::c_void,
+        length: ::std::os::raw::c_int,
     );
 }
 extern "C" {
@@ -3132,6 +3212,15 @@ extern "C" {
     pub fn Fl_Gl_Window_callback(self_: *const Fl_Gl_Window) -> Fl_Callback;
 }
 extern "C" {
+    pub fn Fl_Gl_Window_set_deletion_callback(
+        self_: *mut Fl_Gl_Window,
+        arg1: ::std::option::Option<
+            unsafe extern "C" fn(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
     pub fn Fl_Gl_Window_begin(self_: *mut Fl_Gl_Window);
 }
 extern "C" {
@@ -3327,6 +3416,13 @@ extern "C" {
     pub fn Fl_Gl_Window_set_icon_label(
         arg1: *mut Fl_Gl_Window,
         arg2: *const ::std::os::raw::c_char,
+    );
+}
+extern "C" {
+    pub fn Fl_Gl_Window_set_icons(
+        w: *mut Fl_Gl_Window,
+        images: *mut *const ::std::os::raw::c_void,
+        length: ::std::os::raw::c_int,
     );
 }
 extern "C" {
@@ -3710,6 +3806,15 @@ extern "C" {
     pub fn Fl_Glut_Window_callback(self_: *const Fl_Glut_Window) -> Fl_Callback;
 }
 extern "C" {
+    pub fn Fl_Glut_Window_set_deletion_callback(
+        self_: *mut Fl_Glut_Window,
+        arg1: ::std::option::Option<
+            unsafe extern "C" fn(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
     pub fn Fl_Glut_Window_begin(self_: *mut Fl_Glut_Window);
 }
 extern "C" {
@@ -3912,6 +4017,13 @@ extern "C" {
     pub fn Fl_Glut_Window_set_icon_label(
         arg1: *mut Fl_Glut_Window,
         arg2: *const ::std::os::raw::c_char,
+    );
+}
+extern "C" {
+    pub fn Fl_Glut_Window_set_icons(
+        w: *mut Fl_Glut_Window,
+        images: *mut *const ::std::os::raw::c_void,
+        length: ::std::os::raw::c_int,
     );
 }
 extern "C" {